refactor(cli): migrate tropycal tooling onto discovery
Move tropycal's SDK prober and its SDK-universe validator to earthlens.tropycal.cli.
Relocate the CLI tests. Shrink the guard allow-list (tropycal).
