+++ libs/core/tests/test_distribution_boundaries.py
@@ -1274,7 +1274,6 @@ class TestCliIsBackendAgnostic:
             "ecmwf",
             "nwp",
             "openaq",
-            "tropycal",
         }
     )
